move map identify code to core
